backend/requirements-production.txt
# Production Requirements for DocGen Backend
# Python 3.9+

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo[srv]==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.2
email-validator==2.1.0
authlib==1.3.0

# AI & ML Services
google-generativeai==0.7.2
# Consider migrating to google-genai when stable
# google-genai==0.1.0
huggingface-hub==0.20.3
httpx==0.27.0

# PDF Generation
reportlab==4.0.9
Pillow==10.2.0
PyPDF2==3.0.1

# Image Processing
cairosvg==2.7.1

# Data Visualization
matplotlib==3.8.2
numpy==1.26.3

# File Handling
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Production Monitoring (optional but recommended)
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Rate Limiting
slowapi==0.1.9

# Redis for caching (optional)
# redis==5.0.1
# aioredis==2.0.1

# Background Tasks (optional)
# celery==5.3.4
# celery[redis]==5.3.4